3D Tiles转S3M能力提升： 1.支持以倾斜摄影三维模型、城市精模数据为源数据生成的3D Tiles 1.1转换为S3M数据 2.支持直接挂接gltf格式的3D Tiles数据 3.支持解析draco压缩类型的3D Tiles数据 4.修复混合文件(*.cmpt)会出现挂接错误的问题

+++ ConvertTool/Latest(行标）/S3MBToolkit/S3MBToolkit/Src/I3STranslation.cpp
@@ -1958,7 +1958,7 @@ namespace S3MB
 					{
 						for (unsigned int i = 0; i < nCount; i++)
 						{
-							long lValue = 0;
+                            long long lValue = 0;
 							stream >> lValue;
 
 							Variant varValue(lValue);
