+++ S3M_SDK/S3M_SDK_Lightweight package/Latest(S3M 3.0&行标)/s3m/src/S3MBLayerInfos.cpp
@@ -1023,7 +1023,7 @@ namespace S3MB
 		break;
 		case FT_INT64:
 		{
-			varValue = *(long*)pByte;
+            varValue = *(long long*)pByte;
 		}
 		break;
 		case FT_Double:
